pyproject.toml
[project]
name = "recomendador-peliculas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "optuna>=4.4.0",
    "pandas>=2.3.0",
    "skrub>=0.5.4",
    "tensorflow-cpu>=2.18.0",
]
